Revert "toDouble formatting fix."
This reverts commit 7e3e2dea7afca4aa56df240b6b77bd742e17655d.

+++ lib/src/utils/utilities.dart
@@ -61,7 +61,7 @@ class _Utilities {
           symbol: '',
           decimalDigits: this.settings.fractionDigits,
           locale: 'en_US')
-      .format(amount.toDouble());
+      .format(amount);
 
   /// Returns formatted number with refined separator chars
   String get refineSeparator => baseFormat
